Read description
No signifiicant updates from the last commit, but note that sending is disabled. Sending causes the ESP32 to crash, not sure why. For receiving, it does look like it is receiving data but it is not formatted correctly. Gas Pedal Sensor reading only goes between 47 and 50 ish. May be an issue with integers vs ASCII conversions when implementing TWAI,