Significant Updates, see description
Enabled data transmission on transmitter

Properly receiving some CAN data on receiver

Updated SerialStudio JSON File to properly display data with titles
Note on this above: a delimiter character is needed for the start and end of a frame, so I set both to \n. Not sure if it will allow this since we only have the delimiting char \n at the end of each frame.

Updated Serial Print messages